+++ Assets/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2020-10-13
+
+### Changed
+- Made game event editors more user-friendly.
+- Naming in asset menus for game events.
+- Listeners are now exposed as an unmodifiable `ICollection`.
+
 ## [1.0.0-rc.1] - 2020-10-11
 
 ## [0.7.0] - 2020-10-11